requirements.txt
# Requirements for Blendomatic TUI
# Install these in Blender's Python environment:
# /path/to/blender/python/bin/pip install -r requirements.txt

# Core TUI framework
textual>=0.45.0

# Rich text formatting (dependency of textual)
rich>=13.0.0

# Optional: Enhanced prompts and selections
questionary>=2.0.0

# AWS SDK for worker heartbeat storage (S3)
boto3>=1.34.0

# Environment variable loading
python-dotenv>=1.0.0